+++ calc/src/pack/multiplicacao.java
@@ -1,7 +1,7 @@
 package pack;
 
 public class multiplicacao {
-    public float multiplicar(float x, float b) {
+    public int multiplicar(int x, int b) {
         return x * b; 
     }
 }
